requirements.txt
tensorflow
flask
flask-cors
pillow
numpy
tensorflow==2.17.1
gunicorn
